Create Product of real number
Write a program to find the product of a set of real numbers.

Read the number of values n from the console in the first line. For the next "n" lines, read the real numbers one by one. Print the product of the n numbers on the console as output.

The output value should be rounded off to two decimal places.