Create Average of numbers
Write a program to find the average of n numbers using a while loop. Read an integer, n from the console, which will be the number of values.

Write a while loop to read n numbers from the console and find the average of all these numbers. Print the average value as the output with a precision of 3 decimal places.