Create Multiple of 3 in a range
Write a program to display all the multiples of 3 within the range n and m. Read the value of n and m from the first two lines of the console as input. Print all the multiples of 3 within the range of n and m as output (exclude n and m).