Create Check prime or not
Write a program to check whether the given integer is a prime number or not. Read an integer from the console as input.

If the integer is a prime number, print "n is a prime number", where n is the value of the integer. Else, print "n is not a prime number".