Create Sliding window
You will be given a list of integers, arr, and a single integer k. You must sort the numbers in ascending order and create an array of length k from elements of arr to minimize its unfairness. Call that array arr’. Unfairness of an array is calculated as = max(arr’) – min(arr’)

Where:- max denotes the largest integer in arr’- min denotes the smallest integer in arr’